Moved navbar to separate component
